chore: extract main-specific i18n config
