+++ app/common/configs/i18next.renderer.js
@@ -2,7 +2,7 @@ import i18n from 'i18next';
 import {initReactI18next} from 'react-i18next';
 
 import {i18NextBackend, i18NextBackendOptions} from '../renderer/src/polyfills';
-import {getI18NextOptions} from './app.config';
+import {getI18NextOptions} from './i18next.common';
 
 const i18nextOptions = getI18NextOptions(i18NextBackendOptions);
