+++ app/common/renderer/src/Root.jsx
@@ -2,7 +2,7 @@ import React, {Suspense} from 'react';
 import {Provider} from 'react-redux';
 import {MemoryRouter, Route, Routes} from 'react-router-dom';
 
-import i18n from '../../configs/i18next.config.renderer';
+import i18n from '../../configs/i18next.renderer';
 import Spinner from './components/Spinner/Spinner.jsx';
 import InspectorPage from './containers/InspectorPage';
 import SessionPage from './containers/SessionPage';
