+++ app/common/renderer/src/actions/Session.js
@@ -14,7 +14,7 @@ import moment from 'moment';
 import {v4 as UUID} from 'uuid';
 import {Web2Driver} from 'web2driver';
 
-import i18n from '../../../configs/i18next.config.renderer';
+import i18n from '../../../configs/i18next.renderer';
 import {
   SAVED_SESSIONS,
   SERVER_ARGS,
